clipart/requirements.txt
# Video CTA Overlay Editor — зависимости
# Установка: pip install -r requirements.txt

PyQt6>=6.5.0
moviepy>=1.0.3
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
gitpython>=3.1.30
